requirements.txt
boto3
psycopg2-binary
requests
jsonschema
cfn_resource_provider>=1.0.6
AsIs
